pyproject.toml
[tool.poetry]
name = "price-of-coin"
version = "0.1.0"
description = ""
authors = ["AndrewTarev <153756623+AndrewTarev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
pydantic-settings = "^2.3.4"
gunicorn = "^22.0.0"
greenlet = "^3.0.3"
sqlalchemy = "^2.0.36"
pydantic = "^2.9.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
black = "^24.8.0"
mypy = "^1.11.2"


[tool.poetry.group.http_client.dependencies]
sqlalchemy = "^2.0.36"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
aiohttp = "^3.10.10"
pytest-aioresponses = "^0.2.0"

[tool.isort]
profile = "black"
line_length = 90
multi_line_output = 3
skip_gitignore = true
skip_glob = ["**/alembic/*"]


[tool.mypy]
files = "my_project"
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
pretty = true
strict_optional = true
follow_imports = "silent"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
